feat: Flutter web port — run in browser via same Flask server
## What changed

### Flutter web platform enabled
- flutter create --platforms=web adds android_app/web/ scaffold
- pubspec.yaml: add go_router ^14.0.0

### Platform-compatibility (Android unchanged)
- local_db_service.dart: conditional export → db_mobile.dart (sqflite)
  or db_web.dart (shared_prefs/in-memory stub)
- sync_service.dart: kIsWeb guards around all dart:io file operations;
  new audioUri() helper returns file:// on Android, server URL on web
- api_service.dart: baseUrl() returns '' on web (relative /api/... calls)
- player_screen.dart: use audioUri(), skip local file checks on web
- sentences_screen.dart: same
- settings_screen.dart: kIsWeb guards for path_provider/dart:io usage

### go_router
- main.dart: MaterialApp.router + ShellRoute for proper browser
  back-button and URL routing (/login, /, /review, /lessons, /diary,
  /settings); AuthGate merged into go_router redirect
- login_screen.dart: context.go('/') instead of pushReplacementNamed
- settings_screen.dart: context.go('/login') on logout

### Dockerfile (3-stage build)
- Stage 1: base ML deps (unchanged)
- Stage 2: flutter-web — builds Flutter web release; no Flutter SDK
  in the final image
- Stage 3: app — copies build/web/ from stage 2 into /app/web_build/

### Flask serves Flutter web
- webapp.py: GET / and /<path> serve Flutter web SPA from web_build/;
  API routes (/api/*) are unaffected and take priority

### CI (release.yml)
- New build-flutter-web job: flutter build web --release, uploads zip
- release job: includes lingodiary-web.zip in GitHub Release
- Flutter version fixed to 3.41.9 (matches local + Dockerfile)
