+++ Dockerfile
@@ -41,7 +41,40 @@ RUN pip install --no-cache-dir \
 
 RUN mkdir -p /app/.local && chown -R 1000:1000 /app/.local
 
-# ── Stage 2: app — lightweight deps + source code ────────────────────────────
+# ── Stage 2: flutter-web — build the Flutter web app ─────────────────────────
+# Rebuilt only when the Flutter source or pubspec changes.
+# The output (build/web/) is copied into stage 3 so no Flutter SDK ends up in
+# the final image.
+FROM debian:bookworm-slim AS flutter-web
+
+RUN apt-get update && apt-get install -y --no-install-recommends \
+        curl \
+        git \
+        unzip \
+        xz-utils \
+    && rm -rf /var/lib/apt/lists/*
+
+# Install Flutter (pinned to the same version used in CI)
+ENV FLUTTER_VERSION=3.41.9
+RUN curl -fsSL "https://storage.googleapis.com/flutter_infra_release/releases/stable/linux/flutter_linux_${FLUTTER_VERSION}-stable.tar.xz" \
+    | tar -xJ -C /opt && \
+    /opt/flutter/bin/flutter config --no-analytics && \
+    /opt/flutter/bin/flutter precache --web
+
+ENV PATH="/opt/flutter/bin:$PATH"
+
+WORKDIR /flutter_app
+COPY android_app/pubspec.yaml android_app/pubspec.lock* ./
+RUN flutter pub get
+
+# Copy only the source that affects the web build
+COPY android_app/lib/ ./lib/
+COPY android_app/web/ ./web/
+COPY android_app/assets/ ./assets/ 2>/dev/null || true
+
+RUN flutter build web --release
+
+# ── Stage 3: app — lightweight deps + source code ────────────────────────────
 # Rebuilt on every poetry.lock or source change, but stays fast because
 # the heavy ML packages are already present from stage 1.
 
@@ -71,6 +104,9 @@ RUN poetry config virtualenvs.in-project true && \
 # 🔁 Copy source (invalidates cache only on code changes, not dep changes)
 COPY lingodiary/ /app/lingodiary
 COPY README.md /app/
+
+# ✅ Copy Flutter web build from stage 2
+COPY --from=flutter-web /flutter_app/build/web/ /app/web_build/
 
 # ✅ Install the lingodiary package itself; write lock hash for entrypoint check
 RUN poetry install --only-root && \
