+++ android_app/lib/screens/login_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../services/api_service.dart';
 import '../services/auth_service.dart';
@@ -59,7 +60,7 @@ class _LoginScreenState extends State<LoginScreen> {
       }
       if (mounted) {
         // Replace with main shell
-        Navigator.of(context).pushReplacementNamed('/home');
+        context.go('/');
       }
     } on ApiException catch (e) {
       setState(() => _error = e.message);
