+++ android_app/lib/services/sync_service.dart
@@ -1,10 +1,13 @@
 import 'dart:io';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:path_provider/path_provider.dart';
 import 'api_service.dart';
 
 class SyncService {
   /// Returns the local directory used to cache synced output files.
+  /// Throws UnsupportedError on web (local file system unavailable).
   static Future<Directory> _localOutputDir() async {
+    assert(!kIsWeb, '_localOutputDir must not be called on web');
     final base = await getApplicationDocumentsDirectory();
     final dir = Directory('${base.path}/output');
     await dir.create(recursive: true);
@@ -12,15 +15,33 @@ class SyncService {
   }
 
   /// Returns the full local path for a given relative output file path.
+  /// On web, returns the server URL for that file instead.
   static Future<String> localPath(String relPath) async {
+    if (kIsWeb) {
+      final base = await ApiService.baseUrl();
+      return '$base/api/sync/file/${Uri.encodeFull(relPath)}';
+    }
     final dir = await _localOutputDir();
     return '${dir.path}/$relPath';
   }
 
+  /// Returns a [Uri] suitable for use with just_audio on both platforms.
+  /// On Android: a file:// URI pointing to the locally cached file.
+  /// On web:     an https:// URI pointing to the server's sync endpoint.
+  static Future<Uri> audioUri(String relPath) async {
+    if (kIsWeb) {
+      final base = await ApiService.baseUrl();
+      return Uri.parse('$base/api/sync/file/${Uri.encodeFull(relPath)}');
+    }
+    final path = await localPath(relPath);
+    return Uri.file(path);
+  }
+
   /// Downloads a single file from the server and caches it locally.
+  /// On web, the file is always streamed from the server so this is a no-op.
   /// Returns true on success, false if the server returned a non-200 status.
-  /// Throws on network error.
   static Future<bool> downloadFile(String relPath) async {
+    if (kIsWeb) return true; // web streams directly from server
     final dir = await _localOutputDir();
     final localFile = File('${dir.path}/$relPath');
     final resp = await ApiService.downloadFile(relPath);
@@ -33,13 +54,13 @@ class SyncService {
   }
 
   /// Syncs all output files from the server.
-  /// Downloads files that are missing or have a newer mtime on the server.
-  /// Returns the number of files downloaded.
+  /// On web this is a no-op (audio is always streamed from the server).
   static Future<int> syncFromServer({
     void Function(String message)? onProgress,
     void Function(int current, int total)? onProgressCount,
     bool Function()? isCancelled,
   }) async {
+    if (kIsWeb) return 0;
     final dir = await _localOutputDir();
     final manifest = await ApiService.getSyncManifest();
     final total = manifest.length;
@@ -74,12 +95,14 @@ class SyncService {
   }
 
   /// Syncs files for a single lesson (filtered by base name).
+  /// On web this is a no-op (audio streamed directly from server).
   static Future<int> syncLesson(
     String base, {
     void Function(String message)? onProgress,
     void Function(int current, int total)? onProgressCount,
     bool Function()? isCancelled,
   }) async {
+    if (kIsWeb) return 0;
     final dir = await _localOutputDir();
     final manifest = await ApiService.getLessonManifest(base);
     final total = manifest.length;
@@ -113,8 +136,9 @@ class SyncService {
     return downloaded;
   }
 
-  /// Lists locally cached lesson mp3 files.
+  /// Lists locally cached lesson mp3 files. Returns empty list on web.
   static Future<List<FileSystemEntity>> listLocalMp3s() async {
+    if (kIsWeb) return [];
     final dir = await _localOutputDir();
     if (!await dir.exists()) return [];
     return dir
